Extensive updates to splnr_targets.R
Updates were generally to ensure that the functions could use SpatRaster and sf as arguments. Also, the challenge is that with sf all the target info can be in the sf dataframe, whereas it cannot with SpatRaster. I have thus tried to standardise to have a target dataframe, which can be updated with the targets from splnr_targets_byInverseArea() and splnr_targets_byIUCN2() (or other functions).
1. Rewritten splnr_targets_byInverseArea() to take SpatRaster as arugments
2. Included examples of using both sf and SpatRaster as inputs for splnr_targets_byInverseArea
3. I could not get splnr_targets_byIUCN(), so created splnr_targets_byIUCN2(), which now is simpler and works with sf and SpatRaster.
4. Included examples of using both sf and SpatRaster as inputs for splnr_targets_byIUCN2()
5. Have included the new function combine_IUCN_InverseArea_targets() that combines the df outputs from splnr_targets_byIUCN2() and splnr_targets_byInverseArea()